Module prover_lab::benchmark

Structs

Benchmark
Represents a benchmark.

BenchmarkData
A data entry of a benchmark.

Display
A wrapper around a string for text display.

Constants

COLOR_PALETTE

DARK_GRAY

GRAY

GRAY_PALETTE

LIGHT_GRAY

MEDIUM_GRAY

Functions

benchmark

plot_benchmarks
Plot a set of benchmarks in JupyterLab. TODO: we should pull this apart into multiple functions and also let the plotting work against different backends. This is just a first experiment to see what is possible.

read_benchmark
Read benchmark from data file.

stats_benchmarks
Print statistics for the given set of benchmarks. TODO: would be nice to have a histogram instead of textual output.